More responsive process kill.  Cleanup of source directory to avoid svn errors.
Former-commit-id: a0de8ed95cbe29e2696c59dafa381d749356926f

+++ pwiz_tools/Skyline/SkylineTester/ProcessUtilities.cs
@@ -17,59 +17,46 @@
  * limitations under the License.
  */
 using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using System.Linq;
 using System.Management;
 
 namespace SkylineTester
 {
     static class ProcessUtilities
     {
-        private class ProcessParent
-        {
-            public Process Process;
-            public int ParentId;
-        }
-
         public static void KillProcessTree(Process process)
         {
-            var parentMap = new List<ProcessParent>();
-            var allProcesses = Process.GetProcesses();
-            foreach (var p in allProcesses)
+            int id;
+            try
+            {
+                id = process.Id;
+                process.Kill();
+            }
+            catch (Exception)
+            {
+                // ReSharper disable once RedundantJumpStatement
+                return;
+            }
+
+            KillChild(id, "bjam");
+            KillChild(id, "bsdtar");
+        }
+
+        private static void KillChild(int parentId, string processName)
+        {
+            foreach (var process in Process.GetProcessesByName(processName))
             {
                 try
                 {
-                    var mo = new ManagementObject("win32_process.handle='" + p.Id + "'");
+                    var mo = new ManagementObject("win32_process.handle='" + process.Id + "'");
                     mo.Get();
-                    var parentId = Convert.ToInt32(mo["ParentProcessId"]);
-                    parentMap.Add(new ProcessParent {Process = p, ParentId = parentId});
+                    if (parentId == Convert.ToInt32(mo["ParentProcessId"]))
+                        process.Kill();
                 }
 // ReSharper disable once EmptyGeneralCatchClause
                 catch
                 {
                 }
-            }
-
-            KillProcessTree(process, parentMap);
-        }
-
-        private static void KillProcessTree(Process process, IList<ProcessParent> parentMap)
-        {
-            var id = process.Id;
-
-            try
-            {
-                process.Kill();
-            }
-            catch (Exception)
-            {
-                return;
-            }
-
-            foreach (var processParent in parentMap.Where(p => p.ParentId == id))
-            {
-                KillProcessTree(processParent.Process, parentMap);
             }
         }
     }
